fixed admin controller get all elements list, and added add new entity

+++ src/main/java/com/lgi/theweschshop/shopdata/response/AdminElementListResponse.java
@@ -9,6 +9,6 @@ public class AdminElementListResponse {
 
     public List<AdminElementResponse> elementList;
 
-    public Integer itemsTotal;
+    public Long itemsTotal;
 
 }
